armor table update
after consulting the other published full rulebooks, the "increased by 1" armor table is probably incorrect. this update fixes the table in the CSRD and CSPG to this line of thinking, and moves the "increased by 1" clause to a post-table sidebar -- allowing GMs to use it as they wish. avoided using the word "experienced", though, since it appears in a relevant ability.